+++ examples/utils/avatar/avatarTemplate.tsx
