+++ examples/utils/vehicle/vehicleTemplate.tsx
